+++ src/heuristics/index.ts
@@ -1,3 +1,5 @@
-export * from './diagonal';
-export * from './manhattan';
+import { manhattan } from './manhattan';
+import { diagonal } from './diagonal';
+
+export { manhattan, diagonal };
 export * from './types';