+++ src/types.ts
@@ -1,7 +1,7 @@
 import type * as heuristics from '@/heuristics';
 
 export type Grid = number[][];
-export type Vector2D = [number, number];
+export type Vector = [number, number];
 
 export interface Score {
     g: number;
@@ -9,22 +9,27 @@ export interface Score {
     f: number;
 }
 
-export type OpenTile = [Vector2D, Score, OpenTile|null];
+export type OpenTile = [Vector, Score, OpenTile|null];
 
 export interface ScoreOptions {
-    current: Vector2D;
+    current: Vector;
     parent: OpenTile;
-    goal: Vector2D;
+    goal: Vector;
     heuristic: keyof typeof heuristics;
 }
 
 export interface SearchOptions {
-    from: Vector2D;
-    to: Vector2D;
+    from: Vector;
+    to: Vector;
     grid: Grid;
     heuristic?: keyof typeof heuristics;
     diagonal?: boolean;
     cutCorners?: boolean;
+    /**
+     * When defing a grid with elevation, this is the max distance
+     * that can be stepped up/doown. Defaults to `1`.
+     */
+    stepHeight?: number;
 }
 
 /**
@@ -35,4 +40,4 @@ export interface SearchOptions {
  * Removes the need to calculate this later on if `cutCorners` is
  * disabled.
  */
-export type Neighbor = [Vector2D, [Vector2D, Vector2D]|null];
+export type Neighbor = [Vector, [Vector, Vector]|null];